feat(solana): refactor in order to support Versioned Message
