+++ src/Solana/Signature.cpp
@@ -1,0 +1,13 @@
+// Copyright © 2017-2023 Trust Wallet.
+//
+// This file is part of Trust. The full Trust copyright notice, including
+// terms governing use, modification, and redistribution, is contained in the
+// file LICENSE at the root of the source code distribution tree.
+
+#include "Solana/Signature.h"
+
+namespace TW::Solana {
+    bool Signature::operator==(const Signature& v) const {
+        return bytes == v.bytes;
+    }
+}
